// src/model/UserItem.model.ts
export interface UserItem {
    id?: string;
    username?: string;
    email?: string;
    name?: string;
    address?: string;
    phone?: string;
    profile_picture?: string;
    created_at?: string;
    created_by?: string;
    updated_at?: string;
    updated_by?: string;
    deleted_at?: string;
    message?: string;
}